+++ src/SFA.DAS.ApprenticeCommitments/Application/Commands/CreateRegistrationCommand/CreateRegistrationCommand.cs
@@ -28,8 +28,8 @@ namespace SFA.DAS.ApprenticeCommitments.Application.Commands.CreateRegistrationC
         public DateTime PlannedEndDate { get; set; }
         public DateTime? EmploymentEndDate { get; set; }
         public bool? RecognisePriorLearning { get; set; }
-        public int? DurationReducedByHours { get; set; }
-        public int? DurationReducedBy { get; set; }
+        public short? DurationReducedByHours { get; set; }
+        public short? DurationReducedBy { get; set; }
 
     }
 }
